// src/apispec/src/models/CloneVolumeFilesRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CloneVolumeFileEntry } from './CloneVolumeFileEntry';
import {
    CloneVolumeFileEntryFromJSON,
    CloneVolumeFileEntryFromJSONTyped,
    CloneVolumeFileEntryToJSON,
    CloneVolumeFileEntryToJSONTyped,
} from './CloneVolumeFileEntry';

/**
 * 
 * @export
 * @interface CloneVolumeFilesRequest
 */
export interface CloneVolumeFilesRequest {
    /**
     * Clone strategy. cow_or_copy prefers copy-on-write and falls back to server-side copy.
     * @type {string}
     * @memberof CloneVolumeFilesRequest
     */
    mode?: CloneVolumeFilesRequestModeEnum;
    /**
     * Batch atomicity flag. Only true is currently supported.
     * @type {boolean}
     * @memberof CloneVolumeFilesRequest
     */
    atomic?: boolean;
    /**
     * 
     * @type {Array<CloneVolumeFileEntry>}
     * @memberof CloneVolumeFilesRequest
     */
    entries: Array<CloneVolumeFileEntry>;
}


/**
 * @export
 */
export const CloneVolumeFilesRequestModeEnum = {
    CowOrCopy: 'cow_or_copy',
    CowRequired: 'cow_required',
    Copy: 'copy'
} as const;
export type CloneVolumeFilesRequestModeEnum = typeof CloneVolumeFilesRequestModeEnum[keyof typeof CloneVolumeFilesRequestModeEnum];


/**
 * Check if a given object implements the CloneVolumeFilesRequest interface.
 */
export function instanceOfCloneVolumeFilesRequest(value: object): value is CloneVolumeFilesRequest {
    if (!('entries' in value) || value['entries'] === undefined) return false;
    return true;
}

export function CloneVolumeFilesRequestFromJSON(json: any): CloneVolumeFilesRequest {
    return CloneVolumeFilesRequestFromJSONTyped(json, false);
}

export function CloneVolumeFilesRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CloneVolumeFilesRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'mode': json['mode'] == null ? undefined : json['mode'],
        'atomic': json['atomic'] == null ? undefined : json['atomic'],
        'entries': ((json['entries'] as Array<any>).map(CloneVolumeFileEntryFromJSON)),
    };
}

export function CloneVolumeFilesRequestToJSON(json: any): CloneVolumeFilesRequest {
    return CloneVolumeFilesRequestToJSONTyped(json, false);
}

export function CloneVolumeFilesRequestToJSONTyped(value?: CloneVolumeFilesRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'mode': value['mode'],
        'atomic': value['atomic'],
        'entries': ((value['entries'] as Array<any>).map(CloneVolumeFileEntryToJSON)),
    };
}