// src/apispec/src/models/FileInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by internal-gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FileInfo
 */
export interface FileInfo {
    /**
     * 
     * @type {string}
     * @memberof FileInfo
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof FileInfo
     */
    path?: string;
    /**
     * 
     * @type {string}
     * @memberof FileInfo
     */
    type?: FileInfoTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof FileInfo
     */
    size?: number;
    /**
     * 
     * @type {string}
     * @memberof FileInfo
     */
    mode?: string;
    /**
     * 
     * @type {Date}
     * @memberof FileInfo
     */
    modTime?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof FileInfo
     */
    isLink?: boolean;
    /**
     * 
     * @type {string}
     * @memberof FileInfo
     */
    linkTarget?: string;
}


/**
 * @export
 */
export const FileInfoTypeEnum = {
    File: 'file',
    Dir: 'dir',
    Symlink: 'symlink'
} as const;
export type FileInfoTypeEnum = typeof FileInfoTypeEnum[keyof typeof FileInfoTypeEnum];


/**
 * Check if a given object implements the FileInfo interface.
 */
export function instanceOfFileInfo(value: object): value is FileInfo {
    return true;
}

export function FileInfoFromJSON(json: any): FileInfo {
    return FileInfoFromJSONTyped(json, false);
}

export function FileInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): FileInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'path': json['path'] == null ? undefined : json['path'],
        'type': json['type'] == null ? undefined : json['type'],
        'size': json['size'] == null ? undefined : json['size'],
        'mode': json['mode'] == null ? undefined : json['mode'],
        'modTime': json['mod_time'] == null ? undefined : (new Date(json['mod_time'])),
        'isLink': json['is_link'] == null ? undefined : json['is_link'],
        'linkTarget': json['link_target'] == null ? undefined : json['link_target'],
    };
}

export function FileInfoToJSON(json: any): FileInfo {
    return FileInfoToJSONTyped(json, false);
}

export function FileInfoToJSONTyped(value?: FileInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'path': value['path'],
        'type': value['type'],
        'size': value['size'],
        'mode': value['mode'],
        'mod_time': value['modTime'] == null ? value['modTime'] : value['modTime'].toISOString(),
        'is_link': value['isLink'],
        'link_target': value['linkTarget'],
    };
}